+++ SERPRO/Kotlin/AluGames/src/main/kotlin/modelo/jogo/InfoJogoJson.kt
@@ -1,7 +1,9 @@
 package modelo.jogo
 
+import java.math.BigDecimal
+
 data class InfoJogoJson(
     val titulo: String,
     val capa: String,
-    val preco: Double,
+    val preco: BigDecimal,
     val descricao: String)
